requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.12
faiss-cpu>=1.13.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.29.0
tiktoken>=0.5.0
numpy>=1.21.0
pandas>=2.0.0
sentence-transformers>=2.2.2
